Support CentOS 7 (and under):
CentOS 7 can carry a Perl 5.10.1 without ExtUtils::Manifest or
ExtUtils::Install, even though they are core.

This means that the library will not build by default. Adding these
would have been enough, except they need to be under build/configure
and that was only added in EUMM 6.52, so we're adding that as the
minimum version as well.