Add national formatting of numbers.
This adds a new formatter, 'National', that outputs the number in national
format using the stored nationalPrefixFormattingRule to work out what to
display.

It also adds to StubCountry a format_for_country function; supplying a country
code (text or number) will output the number in national format if it is in
that county, international format otherwise.