Add failing tests for Argentina (AR)
The national prefix transform rule is always applied, even if the
national prefix regex does not capture the necessary groups.

For Argentinian numbers, this causes the "9" prefix to be added to the
international numbers for both fixed and mobile national numbers.  It
should only be applied to mobile numbers.

Mobile phone numbers in Argentina are assigned the same geographic area
codes as fixed lines.  Calling a mobile phone from a fixed line requires
dialling 15, followed by the subscriber's number.  To place a domestic
long distance call from a fixed line to a mobile, trunk code 0 must be
dialled, followed by the area code, then the mobile prefix 15, and
finally the subscriber number.

Mobile prefix 15 and trunk code 0 can be omitted when calling a mobile
from another mobile within Argentina.

The international format for fixed line numbers requires the trunk code
0 to be removed, so:

  "0 3715 65 4320" becomes "+54 3715 65 4320"

The international format for mobile numbers requires the trunk code 0 to
be removed, 9 to be placed before the area code, and the mobile prefix
15 to be removed, so:

  "0 3715 15 65 4320" becomes "+54 9 3715 65 4320"