Fix application of national prefix transform rule
libphonenumber only applied the national prefix transform rule if the
last possible capture group of the national prefix rule regex matches.
That is, it won't apply the transform rule if there are no prefixes to
transform.  Where there are no suitable captures, it will simply delete
the matching prefix.  See:

  com.google.i18n.phonenumbers.maybeStripNationalPrefixAndCarrierCode()

The Number::Phone country stub files, however, always apply the
transform rule.  This can lead to cases where phone numbers are
incorrectly converted to the wrong type.  For example, the international
format for Argentinian mobile numbers requires the trunk code 0 to be
removed, 9 to be placed before the area code, and the mobile prefix 15
to be removed, so:

  "0 3715 15 65 4320" becomes "+54 9 3715 65 4320"

But Number::Phone incorrectly does the same for fixed line numbers:

  "0 3715 65 4320" => "+54 9 3715 65 4320"

It should be:

  "0 3715 65 4320" => "+54 3715 65 4320"

This commit makes the Number::Phone country stub classes more closely
follow the logic in libphonenumber.