skip is_drama tests more thoroughly when mocking
The test for +44 7700 ... was getting around the mocking, because that's
recognised as being a +447700 JE number, not a +44 UK number, and so
loading the full-fat module. Yes, I know, I should really fix the mocking.

There's other issues too though, as not only does libphonenumber not have
is_drama, it also doesn't recognise +44 7700 900 at all, so I feel justified
in not fixing mocking that I wouldn't be able to use anyway!