Be able to validate Denmark's mobile numbers
Our app has an issue with Danish numbers, which cannot be validated as
either fixed line or mobile. This is a minor inconvenience to our users.

The problem is that upstream has fixed lines and mobiles lines defined
as the same regexp and Number::Phone strips the checks for is_mobile and
is_fixed when these are the same, as found in build-data.stubs and
build-tests.pl, introduced by 02f12d42294654dcac700113fef258c6e1a4567c.

I've created an upstream PR to improve the Danish number checks so we in
Perl are going to be aware what Denmark's mobile and landline number
are. I've ran build-data.sh --force with the data from the PR at Google and
the tests in t/stubs.t work as intended.

Only merge after https://github.com/google/libphonenumber/pull/3417 has
been merged into the Google code base.

Signed-off-by: Wesley Schwengle <waterkip@cpan.org>